Rename vtsearch/models/progress.py to labeling_progress.py
The deferred "merge with vtsearch/concurrency/progress.py" from task #6
was based on the assumption that both files were progress-tracking
modules.  On inspection they share nothing but the filename:

  * vtsearch/concurrency/progress.py is the long-running-operation
    infrastructure (ProgressTracker, LoadingTasksTracker, the
    dataset/sort/eval/find singletons, update_progress/get_progress).
  * vtsearch/models/progress.py is a labeling-session analyzer that
    caches per-step trained MLPs and computes stopping-condition
    metrics (analyze_labeling_progress, compute_labeling_status,
    calculate_error_cost_over_time, prediction stability, etc.).

Merging them would have produced an awkward mega-module.  Renaming the
second one to labeling_progress.py removes the name collision, makes
the intent obvious at import sites, and unblocks the eventual
vtsearch/detectors/ carve-out by giving the file a name that won't
conflict with anything when it moves package.

This is a pure rename: import-site updates, no behaviour change.
CLAUDE.md, docs/ML.md, docs/ARCHITECTURE.md, docs/plans/extract-library.md,
and docs/plans/codebase-reorg.md updated to match.
