+++ CLAUDE.md
@@ -117,7 +117,7 @@ If a failure is genuinely outside the scope of the current task (e.g. a flaky ne
 - `vtsearch/state/` — Multi-dataset / multi-detector global state. `core.py` owns `DatasetContext`, `DetectorContext`, `_state_lock`, the `medias`/votes proxy objects, and the context registries. Sub-modules split the operations by concern: `votes.py` (toggle_vote / apply_label / clear_votes), `clicks.py` (vote click times), `processors.py` (autorun extractor/localizer CRUD), `diversity.py` (diversity tree), `media_lookup.py` (origin-keyed lookup, collapse_duplicates). The package `__init__.py` re-exports every public name and hosts a handful of cross-cutting helpers (`snapshot_medias`, `clear_medias`, `clear_all`, `get_inclusion`/`set_inclusion`, dataset-display-name and safe-thresholds accessors)
 - `vtsearch/plugins/` — `PluginRegistry`, `PluginBase`, `PluginField`, and the sentinel-based auto-discovery scanner shared by every plugin family (dataset importers, exporters, label importers/sources, processor importers, settings importers/exporters/sources)
 - `vtsearch/sync/` — Generic `SyncSource[LoadT, SaveT]` base class (built on `PluginBase`) shared by `SettingsSource` and `LabelsetSource` for bidirectional load/save
-- `vtsearch/concurrency/` — Async job manager (`async_jobs.py` — `AsyncJob`, `JobManager`, `eval_jobs`, `learned_sort_jobs`), memory-aware worker capping (`memory_budget.py` — `cap_workers_by_memory`), and dataset/sort/eval progress trackers (`progress.py` — `update_progress`/`get_progress` + sort/eval variants, `cancel_dataset_progress`, `check_dataset_cancelled`). A second progress module still lives under `vtsearch/models/progress.py` (model-cache + loading-task tracker) and is queued for a follow-up merge with this one
+- `vtsearch/concurrency/` — Async job manager (`async_jobs.py` — `AsyncJob`, `JobManager`, `eval_jobs`, `learned_sort_jobs`), memory-aware worker capping (`memory_budget.py` — `cap_workers_by_memory`), and the long-running-operation progress infrastructure (`progress.py` — `ProgressTracker`, `LoadingTasksTracker`, `update_progress`/`get_progress` + sort/eval/find variants, `cancel_dataset_progress`, `check_dataset_cancelled`). The labeling-session analyzer in `vtsearch/models/labeling_progress.py` is a separate, unrelated module that caches per-step trained MLPs and computes stopping-condition metrics — the historical name collision with this package was resolved by renaming `models/progress.py`
 - `vtsearch/security/` — Path / URL / pickle safety: `path_validation.py` (validate_server_filepath, sanitize_template_value, rglob_follow_symlinks, glob_top_level), `url_validation.py` (SSRF guard for outbound HTTP), `pickle.py` (safe_pickle_load + peek_pickle_dataset_summary for dataset pickles)
 - `vtsearch/utils/` — Leftover helpers that don't fit the packages above: `hits.py` (`build_media_hit` — the single source of truth for the scored-media hit dict shared by CLI autodetect and `/api/labels/fill-from-sort`), and `synthetic/` (offline media synthesis — `images.py`, `audio.py`, `video.py` for the SyntheticDatasetImporter)
 - `vtsearch/settings_factory.py` — Accessor factories (`make_accessors`, `make_per_side_setting`, `clamp`, `one_of`) used by `vtsearch/settings.py` to generate get/set pairs from the `_SETTING_SPECS` table
