+++ tests/conftest.py
@@ -98,7 +98,7 @@ from vtsearch.media.audio.audio_generator import GENERATOR_SAMPLE_RATE
 from tests.fixtures.medias import NUM_MEDIAS, init_medias
 from vtsearch.media.audio.audio_generator import generate_wav
 from vtsearch.models import initialize_models, train_and_score
-from vtsearch.models.progress import clear_progress_cache
+from vtsearch.models.labeling_progress import clear_progress_cache
 from vtsearch.state import (
     bad_votes,
     medias,
