+++ tests/core/test_votes.py
@@ -1,5 +1,5 @@
 import app as app_module
-from vtsearch.models.progress import (
+from vtsearch.models.labeling_progress import (
     _cache_good_ids,
     _cache_bad_ids,
     _cached_steps,
@@ -535,7 +535,7 @@ class TestStabilitySkipsUnchangedModel:
 
     def test_unchanged_training_data_skips_stability(self, client):
         """When good/bad IDs don't change between steps, stability should be None."""
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
 
         clear_progress_cache()
 
@@ -566,7 +566,7 @@ class TestStabilitySkipsUnchangedModel:
 
     def test_changed_training_data_records_stability(self, client):
         """When good/bad IDs DO change, stability should be computed (once a prior model exists)."""
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
 
         clear_progress_cache()
 
@@ -608,7 +608,7 @@ class TestLiveModelReuse:
         import numpy as np
         import torch
 
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
         from vtsearch.models.training import train_model
 
         clear_progress_cache()
@@ -649,7 +649,7 @@ class TestLiveModelReuse:
         import numpy as np
         import torch
 
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
         from vtsearch.models.training import train_model
 
         clear_progress_cache()
@@ -686,7 +686,7 @@ class TestLiveModelReuse:
         import numpy as np
         import torch
 
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
         from vtsearch.models.training import train_model
 
         rng = np.random.RandomState(44)
@@ -702,7 +702,7 @@ class TestLiveModelReuse:
 
     def test_learned_sort_injects_model(self, client):
         """The /api/learned-sort endpoint should inject the trained model."""
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
 
         clear_progress_cache()
 
@@ -724,7 +724,7 @@ class TestLiveModelReuse:
         import numpy as np
         import torch
 
-        from vtsearch.models.progress import clear_progress_cache
+        from vtsearch.models.labeling_progress import clear_progress_cache
         from vtsearch.models.training import train_model
 
         clear_progress_cache()
